Rename shared lib to dagnode
