+++ airflow-core/pyproject.toml
@@ -226,7 +226,7 @@ exclude = [
 
 [tool.hatch.build.targets.sdist.force-include]
 "../shared/configuration/src/airflow_shared/configuration" = "src/airflow/_shared/configuration"
-"../shared/definitions/src/airflow_shared/definitions" = "src/airflow/_shared/definitions"
+"../shared/dagnode/src/airflow_shared/dagnode" = "src/airflow/_shared/dagnode"
 "../shared/logging/src/airflow_shared/logging" = "src/airflow/_shared/logging"
 "../shared/module_loading/src/airflow_shared/module_loading" = "src/airflow/_shared/module_loading"
 "../shared/observability/src/airflow_shared/observability" = "src/airflow/_shared/observability"
@@ -301,7 +301,7 @@ apache-airflow-devel-common = { workspace = true }
 [tool.airflow]
 shared_distributions = [
     "apache-airflow-shared-configuration",
-    "apache-airflow-shared-definitions",
+    "apache-airflow-shared-dagnode",
     "apache-airflow-shared-logging",
     "apache-airflow-shared-module-loading",
     "apache-airflow-shared-observability",
