+++ airflow-core/src/airflow/_shared/dagnode
@@ -1,0 +1,1 @@
+../../../../shared/dagnode/src/airflow_shared/dagnode
