--- airflow-core/src/airflow/_shared/definitions
@@ -1,1 +1,0 @@
-../../../../shared/definitions/src/airflow_shared/definitions
