+++ shared/dagnode/src/airflow_shared/dagnode/node.py
