--- task-sdk/src/airflow/sdk/_shared/definitions
@@ -1,1 +1,0 @@
-../../../../../shared/definitions/src/airflow_shared/definitions
